+++ CHANGELOG.md
@@ -5,6 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 - Metadata: activity?, clipping?, externalPerson?
 
+## 1.7.0 - 2017-11-03
+### Added
+- Person extractor - find by identifier.
+- Person - HESA id, Scopus id.
+
+### Fixed
+- Person - employee_id presence.
+
 ## 1.6.0 - 2017-10-24
 ### Added
 - ExternalOrganisation.